// templates/dashboard/model1/src/lib/enums/ButtonVariant.ts
export enum ButtonVariant {
    WHITE = 'white',
    WHITE_BORDER = 'white-border',
    WHITE_MUTED = 'white-muted',
    WHITE_BORDER_MUTED = 'white-border-muted',
    PRIMARY = 'primary',
    PRIMARY_SECONDARY = 'primary-secondary',
    GREEN = 'green',
    GREEN_SECONDARY = 'green-secondary',
    YELLOW = 'yellow',
    YELLOW_SECONDARY = 'yellow-secondary',
    RED = 'red',
    RED_SECONDARY = 'red-secondary',
  }
  